requirements.txt
binance
flask
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
dateparser==1.0.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
numpy==1.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-binance==0.7.9
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
service-identity==18.1.0
six==1.15.0
TA-Lib @ file:///C:/Users/brand/Downloads/TA_Lib-0.4.19-cp39-cp39-win_amd64.whl
Twisted @ file:///C:/Users/brand/Downloads/Twisted-20.3.0-cp39-cp39-win_amd64.whl
txaio==20.12.1
tzlocal==2.1
ujson==4.0.1
urllib3==1.26.2
websocket-client==0.57.0
zope.interface==5.2.0
